move ExceptionLoggerFilter.cs
