+++ Slack.Automation/Promact.Erp.Core/Promact.Erp.Core.csproj
@@ -325,6 +325,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ActionFilters\ExceptionLoggerFilter.cs" />
     <Compile Include="ActionFilters\SampleActionFilter.cs" />
     <Compile Include="Controllers\BaseController.cs" />
     <Compile Include="Controllers\Bot.cs" />
